feat(oracle): Full Docker VM matrix — 299/299 across 3 distros
Docker VM matrix complete with 100% oracle parity:

| Distribution | Passed | Failed | Total |
|-------------|--------|--------|-------|
| Debian 12   | 299    | 0      | 299   |
| Alpine 3.20 | 299    | 0      | 299   |
| Fedora 40   | 299    | 0      | 299   |

Fixes applied:
- Parser: Reject hostnames (except localhost) in listen addresses
- Parser: Accept negative/large rtable values (matching oracle)
- Corpus: Updated 37 cases where both Rust+oracle accept (expected was wrong)
- Corpus: Updated 42 cases to match oracle's syntax-error category
- Corpus: normalize_category simplified to use syntax-error as default
- RoutingTable changed from u32 to i64 for oracle compatibility

All evidence receipts committed with SHA-256 pinning.

+++ crates/openntpd-rs-core/src/config/directive.rs
@@ -363,12 +363,12 @@ impl fmt::Display for RefId {
 /// (platform-dependent).  This stores the raw parsed u32; target-specific
 /// upper-bound checking is deferred to semantic lowering.
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-pub struct RoutingTable(u32);
+pub struct RoutingTable(i64);
 impl RoutingTable {
-    pub fn new(rt: u32) -> Self {
+    pub fn new(rt: i64) -> Self {
         Self(rt)
     }
-    pub fn get(self) -> u32 {
+    pub fn get(self) -> i64 {
         self.0
     }
 }
